backend/requirements.txt
asgiref>=3.3.4
Django>=3.2.3
django-cors-headers>=3.7.0
django-model-utils>=4.1.1
django-phonenumber-field>=5.1.0
djangorestframework>=3.12.4
gunicorn>=20.1.0
phonenumbers>=8.12.24
Pillow>=8.2.0
psycopg2>=2.9.1
pytz>=2021.1
sqlparse>=0.4.1
typing-extensions>=3.10.0.0
whitenoise>=5.2.0